% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affecte_une_classe.R
\name{affecte_une_classe}
\alias{affecte_une_classe}
\title{affecte_une_classe}
\usage{
affecte_une_classe(x, seuil)
}
\arguments{
\item{x}{un nombre ou un vecteur de nombres}

\item{seuil}{un objet de classe seuil ou une liste de classe seuil de longueur 1}
}
\value{
la fonction renvoie un vecteur ou un nombre avec les classes de qualité correspondant aux valeurs en entrée
}
\description{
fonction pour affecter une classe de qualité à des résultats
}
\examples{
seuil<-makeSeuils(CdParametre="1335", type_seuil="DCE")[[1]]
affecte_une_classe(0.6, seuil)
test <- sample(seq(0,0.6, 0.1),10, replace=T)
affecte_une_classe(test, seuil)
}
