% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajoute_nom_cdaccreana.R
\name{ajoute_nom_cdaccreana}
\alias{ajoute_nom_cdaccreana}
\title{ajoute_nom_cdaccreana}
\usage{
ajoute_nom_cdaccreana(x, col_cdaccredi = "CdAccreAna")
}
\arguments{
\item{x}{data.frame avec le code accréditation}

\item{col_cdaccredi}{: nom de la colonne avec le code accréditation Valeur par défaut : CdAccreAna}
}
\value{
la fonction le data.frame complété de la colonne RqAna
}
\description{
fonction pour ajouter le nom du code accréditation sandre à une data frame qui contient des codes accréditation SANDRE
}
\examples{
test<-data.frame(CdAccreAna2=c("1", "2"))
ajoute_nom_cdrq(test, col_cdaccredi="CdAccreAna2")
}
