% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajoute_nom_cddifficulteana.R
\name{ajoute_nom_cddifficulteana}
\alias{ajoute_nom_cddifficulteana}
\title{ajoute_nom_cddifficulteana}
\usage{
ajoute_nom_cddifficulteana(x, col_difficulte = "CdDifficulteAna")
}
\arguments{
\item{x}{data.frame avec le code paramètre}

\item{col_difficulte}{: nom de la colonne avec le code paramètre. Valeur par défaut : CdDifficulteAna}
}
\value{
la fonction le data.frame complété de la colonne RqAna
}
\description{
fonction pour ajouter le nom du code difficulté sandre à une data frame
}
\examples{
test<-data.frame(CdDifficulteAna2=c("1", "10"))
ajoute_nom_cdrq(test, col_difficulte="CdDifficulteAna2")
}
