% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajoute_nom_cdinsituana.R
\name{ajoute_nom_cdinsituana}
\alias{ajoute_nom_cdinsituana}
\title{ajoute_nom_cdinsituana}
\usage{
ajoute_nom_cdinsituana(x, col_insitu = "CdInsituAna")
}
\arguments{
\item{x}{data.frame avec le code mesure in situ}

\item{col_insitu}{: nom de la colonne avec le code paramètre. Valeur par défaut : CdInsituAna}
}
\value{
la fonction le data.frame complété de la colonne RqAna
}
\description{
fonction pour ajouter le nom du code mesure insitu sandre à une data frame
}
\examples{
test<-data.frame(CdInsituAna=c("1", "10"))
ajoute_nom_cdinsituana(test, col_insitu="CdInsituAna")
}
