% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajoute_nom_cdreseaumesure.R
\name{ajoute_nom_cdreseaumesure}
\alias{ajoute_nom_cdreseaumesure}
\title{ajoute_nom_cdreseaumesure}
\usage{
ajoute_nom_cdreseaumesure(x, col_cdreseau = "CdRdd")
}
\arguments{
\item{x}{data.frame avec le code réseau}

\item{col_cdreseau}{: nom de la colonne avec le code accréditation Valeur par défaut : CdRdd}
}
\value{
la fonction le data.frame complété de la colonne NomRdd
}
\description{
fonction pour ajouter le nom du code réseau de mesure sandre à une data frame
}
\examples{
test<-data.frame(CdAccreAna2=c("1", "2"))
ajoute_nom_cdreseaumesure(test, col_cdaccredi="CdAccreAna2")
}
