% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajoute_nom_cdrq.R
\name{ajoute_nom_cdrq}
\alias{ajoute_nom_cdrq}
\title{ajoute_nom_cdrq}
\usage{
ajoute_nom_cdrq(x, col_rq = "CdRqAna")
}
\arguments{
\item{x}{data.frame avec le code remarque}

\item{col_rq}{: nom de la colonne avec le code remarque Valeur par défaut : CdRqAna}
}
\value{
la fonction le data.frame complété de la colonne RqAna
}
\description{
fonction pour ajouter le nom du code remarque sandre à une data frame qui contient des codes remarques SANDRE
}
\examples{
test<-data.frame(CdRqAna2=c("1", "10"))
ajoute_nom_cdrq(test, col_rq="CdRqAna2")
}
