% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajoute_nom_cdstatutana.R
\name{ajoute_nom_cdstatutana}
\alias{ajoute_nom_cdstatutana}
\title{ajoute_nom_cdstatutana}
\usage{
ajoute_nom_cdstatutana(x, col_statut = "CdStatutAna")
}
\arguments{
\item{x}{data.frame avec le code statut}

\item{col_statut}{: nom de la colonne avec le code statut Valeur par défaut : CdStatutAna}
}
\value{
la fonction le data.frame complété de la colonne CdStatutAna
}
\description{
fonction pour ajouter le nom du code statut de l'analyse sandre à une data frame
}
\examples{
test<-data.frame(CdStatutAna=c("1", "10"))
ajoute_nom_cdstatutana(test, col_rq="CdStatutAna")
}
