% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajoute_nom_fraction.R
\name{ajoute_nom_fraction}
\alias{ajoute_nom_fraction}
\title{ajoute_nom_fraction}
\usage{
ajoute_nom_fraction(x, col_fraction = "CdFractionAnalysee")
}
\arguments{
\item{x}{data.frame avec le code paramètre}

\item{col_fraction}{: nom de la colonne avec le code paramètre. Valeur par défaut : CdFractionAnalysee}
}
\value{
la fonction le data.frame complété de la colonne NomParametre
}
\description{
fonction pour ajouter le nom du paramètre sandre à une data frame qui contient des codes paramètres SANDRE
}
\examples{
test<-data.frame(CdFractionAnalysee2=c("3", "23"))
ajoute_nom_fraction(test, col_fraction="CdFractionAnalysee2")
}
