% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajoute_nom_intervenant.R
\name{ajoute_nom_intervenant}
\alias{ajoute_nom_intervenant}
\title{ajoute_nom_intervenant}
\usage{
ajoute_nom_intervenant(
  x,
  col_intervenant = "CdProducteur",
  type = "nom",
  nom_output = ""
)
}
\arguments{
\item{x}{data.frame avec le code paramètre}

\item{type}{: type de la colonne renvoyée : soit nom complet si type = nom (valeur par défaut), soit mnemonique de l'intervenant si type = mnemo}

\item{nom_output}{: nom de la colonne de sortie (par défaut le nom de sortie est NomIntervenant si sortie de type nom ou MnIntervenant si type mnemo)}

\item{col_fraction}{: nom de la colonne avec le code paramètre. Valeur par défaut : CdFractionAnalysee}
}
\value{
la fonction le data.frame complété de la colonne NomParametre
}
\description{
fonction pour ajouter le nom du paramètre sandre à une data frame qui contient des codes paramètres SANDRE
}
\examples{
test<-data.frame(CdProducteur=c("25440124300012"))
ajoute_nom_intervenant(test, col_intervenant="CdProducteur")
}
