% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajoute_nom_methode.R
\name{ajoute_nom_methode}
\alias{ajoute_nom_methode}
\title{ajoute_nom_methode}
\usage{
ajoute_nom_methode(x, col_methode = "CdMethode")
}
\arguments{
\item{x}{data.frame avec le code paramètre}

\item{col_methode}{: nom de la colonne avec le code méthode Valeur par défaut : CdMethode}
}
\value{
la fonction le data.frame complété de la colonne RqAna
}
\description{
fonction pour ajouter le nom de la méthode correspondant à un code méthode selon la nomenclature SANDRE
}
\examples{
test<-data.frame(CdMethode=c("712", "405"))
ajoute_nom_methode(test, col_methode="CdMethode")
}
