% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajoute_nom_param.R
\name{ajoute_nom_param}
\alias{ajoute_nom_param}
\title{ajoute_nom_param}
\usage{
ajoute_nom_param(x, col_parametre = "CdParametre", type_nom = "long")
}
\arguments{
\item{x}{data.frame avec le code paramètre}

\item{col_parametre}{: nom de la colonne avec le code paramètre. Valeur par défaut : CdParametre}

\item{type_nom}{: variable pouvant valoir soit "court", soit "long" selon si on veut le nom complet ou abrégé du référentiel SANDRE. Valeur par défaut : "long"}
}
\value{
la fonction le data.frame complété de la colonne NomParametre
}
\description{
fonction pour ajouter le nom du paramètre sandre à une data frame qui contient des codes paramètres SANDRE
}
\examples{
test<-data.frame(CdParametre2=c("1335", "1340"))
ajoute_nom_param(test, col_parametre="CdParametre2")
}
