% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajoute_nom_station.R
\name{ajoute_nom_station}
\alias{ajoute_nom_station}
\title{ajoute_nom_station}
\usage{
ajoute_nom_station(x, col_station = "CdStationMesureEauxSurface")
}
\arguments{
\item{x}{data.frame avec le code station}

\item{col_station}{: nom de la colonne avec le code station Valeur par défaut : CdStationMesureEauxSurface}
}
\value{
la fonction le data.frame complété de la colonne NomParametre
}
\description{
fonction pour ajouter le nom de la station sandre à une data frame qui contient des codes stations SANDRE
}
\examples{
test<-data.frame(CdStationMesureEauxSurface2=c("03174000", "04207500"))
ajoute_nom_station(test, col_station="CdStationMesureEauxSurface2")
}
