% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajoute_nom_support.R
\name{ajoute_nom_support}
\alias{ajoute_nom_support}
\title{ajoute_nom_support}
\usage{
ajoute_nom_support(x, col_support = "CdSupport")
}
\arguments{
\item{x}{data.frame avec le code support}

\item{col_support}{: nom de la colonne avec le code support. Valeur par défaut : CdSupport}
}
\value{
la fonction le data.frame complété de la colonne NomParametre
}
\description{
fonction pour ajouter le nom du support sandre à une data frame qui contient des codes support SANDRE
}
\examples{
test<-data.frame(CdSupport2=c("3", "6"))
ajoute_nom_support(test, col_support="CdSupport2")
}
