% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajoute_nom_unite.R
\name{ajoute_nom_unite}
\alias{ajoute_nom_unite}
\title{ajoute_nom_unite}
\usage{
ajoute_nom_unite(x, col_unite = "CdUniteMesure", type_nom = "symbole")
}
\arguments{
\item{x}{data.frame avec le code paramètre}

\item{col_unite}{: nom de la colonne avec le code unite Valeur par défaut : CdUniteMesure}

\item{type_nom}{: variable pouvant valoir soit "symbole", soit "nom" selon si on veut le symbole ou le libellé de l'unité du référentiel SANDRE. Valeur par défaut : "symbole"}
}
\value{
la fonction le data.frame complété de la colonne NomParametre
}
\description{
fonction pour ajouter le nom de l'unité sandre à une data frame qui contient des codes unités SANDRE
}
\examples{
test<-data.frame(CdUniteMesure2=c("129", "13", "133"))
ajoute_nom_unite(test, col_unite="CdUniteMesure")
}
