% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrondi.R
\name{arrondi}
\alias{arrondi}
\title{arrondi}
\usage{
arrondi(nb, digits = 0)
}
\arguments{
\item{nb}{nombre à arrondir}

\item{digits}{nombre de chiffres significatifs}
}
\value{
l'arrondi du nombre selon les règles usuelles en France
}
\description{
fonction pour calculer un arrondi conforme aux usages en France (les nombres qui se terminent par 0.5 sont arrondis à l'unité supérieure)
}
\examples{
arrondi(8.5)
}
