% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_ref_eqr.R
\docType{data}
\name{base_ref_eqr}
\alias{base_ref_eqr}
\title{base_ref_eqr}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 328 rows and 6 columns.
}
\usage{
data(base_ref_eqr)
}
\description{
data frame contenant les valeurs de référence pour calculer les EQR des indices biologiques IBMR et IBD, conformément à l'arrêté ministériel du 25 janvier 2010.
}
\details{
Le champs TYPEFR correspond au code de la typologie tel qu'usuellement utilisé.
Par exemple : TP16-A correspond à un Très Petit Cours d'eau d'hydroécorégion 16 (Corse) et de HER de niveau 2 n°22 (type A)
De même : G3/19-8 correspond à un grand cours d'eau de l'hydroécorégion 3 (Massif Central Sud) et Exogène des HER 19 ou 8.
}
\examples{
str(data(base_ref_eqr))
}
\keyword{datasets}
