% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcule_formes_azote.R
\name{calcule_formes_azote}
\alias{calcule_formes_azote}
\title{calcule_formes_azote}
\usage{
calcule_formes_azote(
  data,
  CdRqAna = NULL,
  CdParametre = "CdParametre",
  RsAna = "RsAna",
  data_incomplete_stop = TRUE
)
}
\arguments{
\item{CdRqAna}{character indiquant quelle colonne contient le code remarque SANDRE
de l'analyse. Si le paramètre est NULL (par défaut) alors il n'y a pas de code remarque}

\item{CdParametre}{character indiquant quelle colonne contient le code
parametre SANDRE de l'analyse. Par defaut CdParametre}

\item{RsAna}{character indiquant quelle colonne contient le resultat
de l'analyse. Par defaut RsAna}

\item{data_incomplete_stop}{: booleen : si TRUE alors la fonction s'arrête s'il
manque un des paramètres d'entrée. Si FALSE alors le calcule continue}

\item{donnees}{data.frame avec les colonnes CdParametre (code sandre du parametre),
RsAna (résultat d'analyses), CdRqAna (code remarque analyse, optionnel)}
}
\value{
la fonction renvoie un data.frame (en mg N/L) avec les valeurs
d'azote organique, d'azote ammoniacal en N-NH4, de nitrites en N-NO2 et
de nitrates en N-NO3.
Une valeur avec code remarque différent de 1 est prise comme nulle
}
\description{
fonction pour calculer les différentes formes de l'azote
a partir des parametres entrants azote kjeldahl, ammonium (en NH4/L),
nitrites (en NO2/L) et nitrates (en NO3/L)
}
\examples{
data<-data.frame(CdParametre=c("1335", "1339", "1340", "1319"),
                            RsAna=c(0.5,0.1,30,2),
                            CdRqAna=c("1","10","1","1"))
calcule_formes_azote(data)

}
