% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcule_formes_phosphore.R
\name{calcule_formes_phosphore}
\alias{calcule_formes_phosphore}
\title{calcule_formes_phosphore}
\usage{
calcule_formes_phosphore(
  data,
  CdRqAna = NULL,
  CdParametre = "CdParametre",
  RsAna = "RsAna"
)
}
\arguments{
\item{CdRqAna}{character indiquant quelle colonne contient le code remarque SANDRE
de l'analyse. Si le paramètre est NULL (par défaut) alors il n'y a pas de code remarque}

\item{CdParametre}{character indiquant quelle colonne contient le code
parametre SANDRE de l'analyse. Par defaut CdParametre}

\item{RsAna}{character indiquant quelle colonne contient le resultat
de l'analyse. Par defaut RsAna}

\item{donnees}{data.frame avec les colonnes CdParametre (code sandre du parametre),
RsAna (résultat d'analyses), CdRqAna (code remarque analyse, optionnel)}
}
\value{
la fonction renvoie un data.frame (en mg P/L) avec les valeurs
de phosphore sous forme PO4 et des autres formes de phosphores. La ligne autres formes de phosphore
n'exite pas sousle SANDRE. Il est indique Pautre
Une valeur avec code remarque différent de 1 est prise comme nulle
}
\description{
fonction pour calculer les différentes formes de phosphore
a partir des parametres entrants phosphore total et orthophosphates
}
\examples{
data<-data.frame(CdParametre=c("1350", "1433"),
                            RsAna=c(0.6,0.5),
                            CdRqAna=c("1","1"))
calcule_formes_phosphore(data)

}
