% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcule_SEEE_I2M2.R
\name{calcule_SEEE_I2M2}
\alias{calcule_SEEE_I2M2}
\title{calcule_SEEE_I2M2}
\usage{
calcule_SEEE_I2M2(donnees)
}
\arguments{
\item{donnees}{= data.table avec les colonnes requises par le script}
}
\value{
Résultats outil I2M2 invertébrés SEEE
}
\description{
fonction pour traiter une liste I2M2 via le script diagnostic invertébrés du SEEE
voir : https://seee.eaufrance.fr/ outil de diagnostic invertébrés v. 1.0.2
}
\details{
Les colonnes à passer impérativement dans la table d'entrée sont celles prévues dans le script :
\itemize{
\item CODE_OPERATION : identifiant unique de l'opération
\item CODE_STATION : code SANDRE de la station de mesure
\item DATE : date de l'opération (format "\%d/\%m/\%Y")
\item TYPONATIONALE : Typologie nationale de la station (ex. TP12-A)
\item CODE_PHASE : code de la phase de prélèvement (A, B ou C)
\item CODE_TAXON : code SANDRE du taxon
\item RESULTAT : effectif du taxon
\item CODE_REMARQUE : code remarque selon le SANDRE
}
}
\examples{
stations_op<-import_hubeau_indices_hbio(liste_stations = "04207400", indice="inv")
stations_op<-stations_op\%>\%subset(CdParametre=="7613")
donnees<-import_hubeau_liste_hbio(liste_stations = c(station_etudiee), indice="inv")
donnees<-donnees\%>\%subset(date_prelevement\%in\%stations_op$DatePrel)
donnees$CODE_OPERATION <-
paste0(donnees$code_station_hydrobio, "*", donnees$date_prelevement)
donnees$CODE_STATION <- donnees$code_station_hydrobio
donnees$DATE <- format(donnees$date_prelevement, "\%d/\%m/\%Y")
donnees$TYPO_NATIONALE <- "P12-A"
donnees$CODE_PHASE <- donnees$code_lot
donnees$CODE_TAXON <- donnees$code_appel_taxon
donnees$RESULTAT <- donnees$resultat_taxon
donnees$CODE_REMARQUE <- donnees$code_type_resultat
donnees <- donnees \%>\% select(CODE_OPERATION, CODE_STATION, DATE, TYPONATIONALE, CODE_PHASE, CODE_TAXON, RESULTAT, CODE_REMARQUE)
calcule_SEEE_I2M2(donnees)

}
