% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcule_SEEE_IBD.R
\name{calcule_SEEE_IBD}
\alias{calcule_SEEE_IBD}
\title{calcule_SEEE_IBD}
\usage{
calcule_SEEE_IBD(donnees)
}
\arguments{
\item{donnees}{= data.table avec les colonnes requises par le script}
}
\value{
Résultats outil IBD SEEE
}
\description{
fonction pour traiter une liste de diatomées via le script IBD du SEEE
voir : https://seee.eaufrance.fr/ outil d'évaluation IBD v1.2.4
}
\details{
Les colonnes à passer impérativement dans la table d'entrée sont celles prévues dans le script :
\itemize{
\item CODE_OPERATION : identifiant unique de l'opération
\item CODE_STATION : code SANDRE de la station de mesure
\item DATE : date de l'opération (format "\%d/\%m/\%Y")
\item TYPONATIONALE : Typologie nationale de la station (ex. TP12-A)
\item CODE_PHASE : code de la phase de prélèvement (A, B ou C)
\item CODE_TAXON : code SANDRE du taxon
\item RESULTAT : effectif du taxon
\item CODE_REMARQUE : code remarque selon le SANDRE
}
}
\examples{
stations_op<-import_hubeau_indices_hbio(liste_stations = "04207400", indice="dia")
stations_op<-stations_op\%>\%subset(CdParametre=="5856")
donnees<-import_hubeau_liste_hbio(liste_stations = c(station_etudiee), indice="dia")
donnees<-donnees\%>\%subset(date_prelevement\%in\%stations_op$DatePrel)
donnees$CODE_OPERATION <-
paste0(donnees$code_station_hydrobio, "*", donnees$date_prelevement)
donnees$CODE_STATION <- donnees$code_station_hydrobio
donnees$DATE <- format(donnees$date_prelevement, "\%d/\%m/\%Y")
donnees$CODE_TAXON <- donnees$code_appel_taxon
donnees$RESULTAT <- donnees$resultat_taxon
donnees <- donnees \%>\% select(CODE_OPERATION, CODE_STATION, DATE, CODE_TAXON, RESULTAT)
calcule_SEEE_IBD(donnees)

}
