% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcule_SEEE_IBMR.R
\name{calcule_SEEE_IBMR}
\alias{calcule_SEEE_IBMR}
\title{calcule_SEEE_IBMR}
\usage{
calcule_SEEE_IBMR(donnees)
}
\arguments{
\item{donnees}{= data.table avec les colonnes requises par le script}
}
\value{
Résultats outil IBMR SEEE
}
\description{
fonction pour traiter une liste de diatomées via le script IBMR du SEEE
voir : https://seee.eaufrance.fr/ outil d'évaluation IBMR v1.1.5
}
\details{
Les colonnes à passer impérativement dans la table d'entrée sont celles prévues dans le script :
\itemize{
\item CODE_OPERATION : identifiant unique de l'opération
\item CODE_STATION : code SANDRE de la station de mesure
\item DATE : date de l'opération (format "\%d/\%m/\%Y")
\item CODE_TAXON : code SANDRE du taxon
\item RESULTAT : Pourcentage de recouvrement du taxon identifié. Le champ est obligatoire et compris entre 0 et 100
\item UR : Unité de relevé. Le champ est obligatoire et seules les modalités F1, F2 et FU sont acceptées
\item POURCENTAGE_FACIES :  Pourcentage du faciès dominant dans l’unité de relevé. compris entre 0 et 100.
}
}
\examples{
stations_op<-import_hubeau_indices_hbio(liste_stations = "04207400", indice="mphy")
stations_op<-stations_op\%>\%subset(CdParametre=="2928")
donnees<-donnees\%>\%subset(libelle_type_resultat=="RecTax")
donnees$CODE_OPERATION <-paste0(donnees$code_station_hydrobio, "*", donnees$date_prelevement)
donnees$CODE_STATION <- donnees$code_station_hydrobio
donnees$DATE <- format(donnees$date_prelevement, "\%d/\%m/\%Y")
donnees$UR<-donnees$code_lot
donnees$CODE_TAXON <- donnees$code_appel_taxon
donnees$RESULTAT <- donnees$resultat_taxon*100
 # nota : dans Hub Eau les pourcentages de recouvrement de chaque faciès ne sont pas fourni ce qui rend impossible le calcul de l'IBMR
donnees$POURCENTAGE_FACIES <-ifelse(donnees$UR=="FU", 100, 50)
donnees <- donnees \%>\% select(CODE_OPERATION, CODE_STATION, DATE, CODE_TAXON, RESULTAT, UR, POURCENTAGE_FACIES)
donnees<-donnees\%>\%subset(!is.na(CODE_TAXON))
calcule_SEEE_IBMR(donnees)

}
