% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcule_ymaxi.R
\name{calcule_ymaxi}
\alias{calcule_ymaxi}
\title{calcule_ymaxi}
\usage{
calcule_ymaxi(donnees)
}
\arguments{
\item{donnees}{vecteur de données}
}
\value{
la fonction renvoie l'échelle maxi permettant d'afficher un maximum d'information

Cette échelle est calculée selon la règle suivante :
\itemize{
\item s'il y a moins de 10 données, ymaxi = 1.2*max(donnees, na.rm=T)
\item s'il y a plus de 10 données, ymaxi est calculé de manière empirique comme max(4 x interquantile(10\%-90\%) + quartile 90\%)
}
}
\description{
fonction pour afficher les graphiques des paramètres en couleurs (mode point)
}
\examples{
data<-runif(30, 0, 10)\%>\%round(2)
calcule_ymaxi(data)

}
