% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceiling_dec.R
\name{ceiling_dec}
\alias{ceiling_dec}
\title{ceiling_dec}
\usage{
ceiling_dec(x, nb_dec)
}
\arguments{
\item{x}{un nombre}

\item{nb_dec}{nombre de décimales}
}
\value{
la fonction renvoie un vescteur avec les données extrèmes remplacées par ymin ou ymax.
}
\description{
fonction pour arrondir à la valeur supérieure en précisant le nb de décimales de l'arrondi
}
\examples{
ceiling_dec(3.2000,2)
}
