% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charge_shp_ICPE.R
\name{charge_shp_ICPE}
\alias{charge_shp_ICPE}
\title{charge_shp_ICPE}
\usage{
charge_shp_ICPE(crs = 2154, shp_emprise = NULL)
}
\arguments{
\item{crs}{valeur du code de projection dans lequel renvoyer le résultat (par défaut Lambert 93, indiquer 4326 pour du wgs84)}

\item{shp_emprise}{vecteur (optionnel) objet sf dans lequel on va récupérer les ICPE}
}
\value{
la fonction renvoie un objet sf avec les ICPE de l'emprise indiquée
}
\description{
charge sous forme d'objet sf le shape des ICPE depuis https://www.georisques.gouv.fr/donnees/bases-de-donnees/installations-industrielles
}
\examples{
ICPE<-charge_shp_ICPE(nom_sage="Vilaine")
}
