% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charge_shp_SAGE.R
\name{charge_shp_SAGE}
\alias{charge_shp_SAGE}
\title{charge_shp_SAGE}
\usage{
charge_shp_SAGE(nom_sage = NULL, crs = 2154)
}
\arguments{
\item{nom_sage}{vecteur (optionnel) listant les périmètres de SAGE à retenir}

\item{crs}{valeur du code de projection dans lequel renvoyer le résultat (par défaut Lambert 93, indiquer 4326 pour du wgs84)}
}
\value{
la fonction renvoie un objet sf avec l'emprise des SAGE sélectionnés et les attributs correspondants
}
\description{
charge sous forme d'objet sf le shape des SAGE en France (ou un extrait) à partir du flux wms du SANDRE
}
\examples{
Vilaine<-charge_shp_SAGE(nom_sage="Vilaine")
}
