% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charge_shp_STAQ.R
\name{charge_shp_STAQ}
\alias{charge_shp_STAQ}
\title{charge_shp_STAQ}
\usage{
charge_shp_STAQ(crs = 2154, shp_emprise = NULL)
}
\arguments{
\item{crs}{valeur du code de projection dans lequel renvoyer le résultat (par défaut Lambert 93, indiquer 4326 pour du wgs84)}

\item{shp_emprise}{objet SF qui délimite le périmètre sur lequel il faut renvoyer les stations}
}
\value{
la fonction renvoie un objet sf avec l'emprise des SAGE sélectionnés et les attributs correspondants
}
\description{
charge sous forme d'objet sf le shape des stations de qualité en France (ou un extrait) à partir du flux wms du SANDRE
}
\examples{
Vilaine<-charge_shp_SAGE(nom_sage="Vilaine")
stations<-charge_shp_STAQ(shp_emprise=Vilaine)
}
