% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charge_shp_STEP.R
\name{charge_shp_STEP}
\alias{charge_shp_STEP}
\title{charge_shp_STEP}
\usage{
charge_shp_STEP(crs = 2154, shp_emprise = NULL)
}
\arguments{
\item{crs}{valeur du code de projection dans lequel renvoyer le résultat (par défaut Lambert 93, indiquer 4326 pour du wgs84)}

\item{shp_emprise}{objet SF qui délimite le périmètre sur lequel il faut renvoyer les stations}
}
\value{
la fonction renvoie un objet sf des STEP tel que disponible sous l'atlas cartographique du SANDRE
}
\description{
charge sous forme d'objet sf le shape des stations d'épurations en France (ou un extrait) à partir du flux wms du SANDRE
}
\examples{
StepFR<-charge_shp_STEP()
}
