% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compte_decimales.R
\name{compte_decimales}
\alias{compte_decimales}
\title{compte_decimales}
\usage{
compte_decimales(x)
}
\arguments{
\item{x}{un nombre ou un vecteur de nombres}
}
\value{
la fonction renvoie un chiffre indiquant le nombre de décimales
}
\description{
fonction pour calculer le nombre de décimales d'un nombre
}
\examples{
compte_decimales(3.234)
}
