% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/couleurs_classes.R
\docType{data}
\name{couleurs_classes}
\alias{couleurs_classes}
\title{couleurs_classes}
\format{
An object of class \code{data.frame} with 68 rows and 3 columns.
}
\usage{
data(couleurs_classes)
}
\description{
data frame contenant les correspondances seuils - couleur pour la réalisation de graphs qualité
}
\details{
La colonne TYPE renvoi aux types de la base_seuils
La colonne CLASSE renvoie aux classes de la base_seuils
La colonne NOM_COULEUR correspond au code couleur à utiliser pour la classe de qualité du type de seuil correspondant
}
\examples{
head(data(couleurs_classes))
}
\keyword{datasets}
