% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphDCE_bar.R
\name{graphDCE_bar}
\alias{graphDCE_bar}
\title{graphDCE_points}
\usage{
graphDCE_bar(
  data,
  col_annee = "annee",
  col_valeurs = "RsAna",
  ymaxi = NULL,
  auto_ymaxi = TRUE,
  seuils = NULL,
  unite = NULL,
  titre = NULL,
  taille_titre = 12,
  affiche_legende = T,
  nom_legende = NULL,
  taille_legende = 12,
  lignes = NULL,
  alpha = 0.8
)
}
\arguments{
\item{data}{tableau de données (dataframe)}

\item{col_annee}{nom de la colonne qui contient les années. Par défaut : "annee" formats character ou numeric}

\item{col_valeurs}{nom de la colonne qui contient les valeurs d'analyses.Par défaut : "RsAna"}

\item{ymaxi}{force l'échelle des valeurs (si non précisé, échelle automatique de ggplot2)}

\item{auto_ymaxi}{booléen. Si le paramètre vaut TRUE et que ymaxi n'est pas renseigné alors ce dernier est calculé automatiquement de manière à rendre le graph aussi lisible que possible}

\item{seuils}{objet de classe seuil (factultatif)}

\item{unite}{unité du paramètre (par défaut si seuils est renseigné : le label SANDRE correspondant au slot code_unite du paramètre seuils)}

\item{titre}{titre du graphique (si absent et si seuils et renseigné, la légende par défaut sera le slot nom_parametre)}

\item{taille_titre}{taille police du titre (par défaut 12)}

\item{affiche_legende}{booléen indiquant s'il faut afficher la légende (par défaut TRUE)}

\item{nom_legende}{titre de la légende (si absent et si seuils et renseigné, la légende par défaut sera le slot nom_seuil de la légende, si seuil n'est pas renseigné, la légende sera "Légende")}

\item{taille_legende}{taille police caractères de la légende}

\item{lignes}{vecteur permettant d'ajouter des lignes horizontales au graphique. ex c(10, 25)}

\item{alpha}{transparence des applats de couleurs}
}
\value{
la fonction renvoie un graphique de classe ggplot
}
\description{
fonction pour afficher les graphiques des paramètres en couleurs (mode histogramme)
}
\examples{
data<-data.frame(annee=seq(2010,2013), RsAna=c(12,15.5,67,18.3))
graphDCE_bar(data, seuils=makeSeuils(CdParametre = "1340", type_seuil="DCE"))

}
