% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphDCE_boxplot.R
\name{graphDCE_boxplot}
\alias{graphDCE_boxplot}
\title{graphDCE_boxplot}
\usage{
graphDCE_boxplot(
  data,
  col_mois = "mois",
  col_valeurs = "RsAna",
  ymaxi = NULL,
  auto_ymaxi = TRUE,
  seuils = NULL,
  unite = NULL,
  xlab = "",
  titre = NULL,
  taille_titre = 12,
  affiche_legende = T,
  nom_legende = NULL,
  taille_legende = 12,
  lignes = NULL,
  alpha = 0.8
)
}
\arguments{
\item{data}{tableau de données (dataframe)}

\item{col_mois}{nom de la colonne qui contient les mois. Par défaut : "mois" formats character ou numeric}

\item{col_valeurs}{nom de la colonne qui contient les valeurs d'analyses.Par défaut : "RsAna"}

\item{ymaxi}{force l'échelle des valeurs (si non précisé, échelle automatique de ggplot2)}

\item{auto_ymaxi}{booléen. Si le paramètre vaut TRUE et que ymaxi n'est pas renseigné alors ce dernier est calculé automatiquement de manière à rendre le graph aussi lisible que possible}

\item{seuils}{objet de classe seuil (factultatif)}

\item{unite}{unité du paramètre (par défaut si seuils est renseigné : le label SANDRE correspondant au slot code_unite du paramètre seuils)}

\item{xlab}{étiquette de l'axe des abscisses (vide par défaut)}

\item{titre}{titre du graphique (si absent et si seuils et renseigné, la légende par défaut sera distribution suivi de la valeur du  le slot nom_parametre)}

\item{taille_titre}{taille police du titre (par défaut 12)}

\item{affiche_legende}{booléen indiquant s'il faut afficher la légende (par défaut TRUE)}

\item{nom_legende}{titre de la légende (si absent et si seuils et renseigné, la légende par défaut sera le slot nom_seuil de la légende, si seuil n'est pas renseigné, la légende sera "Légende")}

\item{taille_legende}{taille police caractères de la légende}

\item{lignes}{vecteur permettant d'ajouter des lignes horizontales au graphique. ex c(10, 25)}

\item{alpha}{transparence des applats de couleurs}
}
\value{
la fonction renvoie un graphique de classe ggplot
}
\description{
fonction pour afficher les graphiques des paramètres en couleurs (mode boxplot).
Utilisé par exemple pour visualiser les mois avec les plus hautes concentrations.
}
\examples{
data0<-data.frame(DatePrel=Sys.Date() + sort(sample(1:2000, 100)), RsAna=c(round(runif(100,0,100), 0)), LqAna=c(3))
data0$mois<-format(data0$DatePrel, "\%m")\%>\%factor
graphDCE_boxplot(data0, seuils=makeSeuils(CdParametre = "1340", type_seuil="DCE"), affiche_LQ = T, ymini = 0)

}
