% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphDCE_distribution.R
\name{graphDCE_distribution}
\alias{graphDCE_distribution}
\title{graphDCE_distribution}
\usage{
graphDCE_distribution(
  donnees,
  col_parametre = "CdParametre",
  col_Valeur = "RsAna",
  col_CdRq = "CdRqAna",
  seuil
)
}
\arguments{
\item{donnees}{: dataframe contenant les analyses dont il faut faire le bilan}

\item{col_parametre}{: nom de la colonne par rapport à laquelle faire le bilan (par défaut NomParametre)}

\item{col_Valeur}{: nom de la colonne avec les résultats d'analyses (par défaut RsAna)}

\item{col_CdRq}{: nom de la colonne avec le code remarque des analyses (par défaut CdRqAna). Les codes correspondent à ceux définis par le SANDRE.}

\item{seuil}{: object de classe seuil ou liste d'objets de classe seuil}
}
\value{
la fonction renvoie un chiffre indiquant le nombre de décimales
}
\description{
fonction pour réaliser un graphique bilan de quantifications
}
\examples{
compte_decimales(3.234)
}
