% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupe_tableau_distribution.R
\name{groupe_tableau_distribution}
\alias{groupe_tableau_distribution}
\title{groupe_tableau_distribution}
\usage{
groupe_tableau_distribution(
  donnees,
  col_Valeur = "RsAna",
  col_CdRq = "CdRqAna",
  col_support = "CdSupport",
  col_CdFraction = "CdFractionAnalysee",
  seuils
)
}
\arguments{
\item{donnees}{: dataframe contenant les analyses dont il faut faire le bilan}

\item{col_Valeur}{: nom de la colonne avec les résultats d'analyses (par défaut RsAna)}

\item{col_CdRq}{: nom de la colonne avec le code remarque des analyses (par défaut CdRqAna). Les codes correspondent à ceux définis par le SANDRE.}

\item{col_CdFraction}{: nom de la colonne avec les codes SANDRE des fractions analysés (par défaut : CdFractionAnalysee), renseigner NULL si pas de colonne correspondant}

\item{seuils}{: liste d'objets de classe seuil}

\item{col_parametre}{: nom de la colonne par rapport à laquelle faire le bilan (par défaut NomParametre)}

\item{col_CdSupport:}{nom de la colonne avec les codes SANDRE des supports analysés (par défaut : CdSupport), renseigner NULL si pas de colonne correspondant}
}
\value{
la fonction renvoie un chiffre indiquant le nombre de décimales
}
\description{
fonction pour réaliser grouper les tableaux de distribution
}
\examples{
compte_decimales(3.234)
}
