% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_donnees_ARS.R
\name{import_donnees_ARS}
\alias{import_donnees_ARS}
\title{import_donnees_ARS}
\usage{
import_donnees_ARS(x, shp = NULL, type = "eau_brute")
}
\arguments{
\item{x}{un fichier .zip exporté depuis la base nationale Naïades}

\item{shp}{objet sf facultatif. Si le paramètre est renseigné, seules les données acquises à l'intérieur du périmètre sont conservées.}

\item{type=}{type de données : soit "eau_brute" (par défaut : correspond à la ressource en eau brute),soit "traitement" : correspond aux données sur l'eau traitée,  soit "eau_distribuee" : correspond à l'eau distribuée.}
}
\value{
la fonction renvoie une liste de data.frame avec les résultats contenus dans le fichier
}
\description{
fonction pour importer un fichier .zip exporté depuis le site data.gouv.fr sous forme d'objets R
https://www.data.gouv.fr/api/1/datasets/resultats-du-controle-sanitaire-de-leau-du-robinet/
}
\examples{
x<-"http://data.cquest.org/sise-eaux/eaurob-2016.zip"
fichier<-paste0(tempdir(), "\\\\donnees.zip")
download.file(x, destfile=fichier, mode="wb")
import_donnees_ARS(fichier, charge_shp_SAGE("Vilaine"), type="eau_brute")
}
