% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_hubeau_staq_hbio.R
\name{import_hub_eau_staq_hbio}
\alias{import_hub_eau_staq_hbio}
\title{import_hub_eau_staq_hbio}
\usage{
import_hub_eau_staq_hbio(
  emprise = NULL,
  liste_stations = NULL,
  an_debut = 1900,
  an_fin = format(Sys.Date(), "\%Y") \%>\% as.numeric(),
  indice = c("dia", "mphy", "inv", "poi", "oli")
)
}
\arguments{
\item{emprise}{objet sf optionnel décrivant l'emprise des stations à récupérer (sans précision on récupère sur la france entière)}

\item{liste_stations}{vecteur de character listant les codes sandre des stations à exporter. Longueur maximum de ce vecteur : 100 stations (restrictions inhérentes à l'API)}

\item{an_debut}{début de la période d'export (par défaut 1900)}

\item{an_fin}{fin de la période d'export (par défaut l'année en court)}

\item{indice}{vecteur avec la liste des indices à recherche (dia = diatomées, mphy = macrophytes, inv = macroinvertébrés, poi = poisson, oli = oligochètes). Par défaut la totalité des indices est sélectionnée.}
}
\value{
la fonction renvoie un tibble avec la liste des stations de mesures concernées
}
\description{
Importe depuis Hubeau une liste de station avec des indices hydrobiologiques
https://hubeau.eaufrance.fr/page/api-hydrobiologie
}
\details{
Si les paramètres emprise et liste stations sont présents tous les 2, seul le paramètre liste_stations est pris en compte.
}
\examples{
Vilaine<-charge_shp_SAGE(nom_sage="Vilaine")
import_hub_eau_staq_hbio(emprise=Vilaine, an_debut=2010)
import_hub_eau_staq_hbio(liste_stations=c("03174000", "04216000"))
}
