% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_hubeau_data_hydrometrique.R
\name{import_hubeau_data_hydrometrique}
\alias{import_hubeau_data_hydrometrique}
\title{import_hubeau_data_hydrometrique}
\usage{
import_hubeau_data_hydrometrique(
  code_entite,
  date_debut_obs_elab = NULL,
  date_fin_obs_elab = NULL
)
}
\arguments{
\item{code_entite}{character ou vecteur avec le code de la ou des stations hydrométriques dont on souhaite récupérer les données (obligatoire)}

\item{date_debut_obs_elab}{date de début des observations à récupérer (optionnel)}

\item{date_fin_obs_elab}{début de fin des observations à récupérer (optionnel)}
}
\value{
la fonction renvoie un tibble avec le code de la station hydrométrique, la date et le débit journalier moyen en L/s
}
\description{
Importe depuis Hubeau les débits journaliers moyens d'une station
https://hubeau.eaufrance.fr/page/api-hydrometrie#/hydrometrie/observationsElaborees
}
\examples{
Vilaine<-charge_shp_SAGE(nom_sage="Vilaine")
import_hubeau_data_hydrometrique(code_entite="J930061101", date_debut_obs_elab=as.Date("2000-01-01"))
}
