% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_hubeau_liste_hbio.R
\name{import_hubeau_liste_hbio}
\alias{import_hubeau_liste_hbio}
\title{import_hubeau_liste_hbio}
\usage{
import_hubeau_liste_hbio(
  liste_stations = NULL,
  an_debut = 1900,
  an_fin = format(Sys.Date(), "\%Y") \%>\% as.numeric(),
  indice = c("dia", "mphy", "inv", "poi", "oli"),
  alarme = "stop"
)
}
\arguments{
\item{liste_stations}{vecteur de character listant les codes sandre des stations à exporter. Longueur maximum de ce vecteur : 100 stations (restrictions inhérentes à l'API)}

\item{an_debut}{début de la période d'export (par défaut 1900)}

\item{an_fin}{fin de la période d'export (par défaut l'année en court)}

\item{indice}{vecteur avec la liste des indices à recherche (dia = diatomées, mphy = macrophytes, inv = macroinvertébrés, poi = poisson, oli = oligochètes). Par défaut la totalité des indices est sélectionnée.}

\item{alarme}{Comportement à avoir si la requête renvoi un code erreur : "stop" : renvoi une erreur bloquante "warning" : envoi un simple message d'alerte}
}
\value{
la fonction renvoie un tibble avec la liste des stations de mesures concernées
}
\description{
Importe depuis Hubeau de listes faunistiques ou floristiques à partir de l'API Hubeau
https://hubeau.eaufrance.fr/page/api-hydrobiologie
}
\examples{
import_hubeau_liste_hbio(liste_stations=c("03174000", "04216000"))
}
