% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_Naiades_PC.R
\name{import_Naiades_PC}
\alias{import_Naiades_PC}
\title{import_Naiades_PC}
\usage{
import_Naiades_PC(x, shp = NULL)
}
\arguments{
\item{x}{un fichier .zip exporté depuis la base nationale Naïades}

\item{shp}{objet sf factultatif. Si le paramètre est renseigné, seules les données acquises à l'intérieur du périmètre sont conservées.}
}
\value{
la fonction renvoie une liste de data.frame avec les résultats contenus dans le fichier
}
\description{
fonction pour importer un fichier .zip exporté de Naiades sous forme d'objets R
}
\examples{
x<-"C:\\\\Users\\\\anthony.deburghrave\\\\OneDrive - EPTB Vilaine\\\\Documents\\\\R_Anthony\\\\Naiades\\\\exports_naiades\\\\naiades_export_1950_1989_au 1er decembre 2020.zip"
import_Naiades_PC(x, charge_shp_SAGE("Vilaine"))
}
