% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_QESU_PHY_v3.R
\name{import_QESU_PHY_v3}
\alias{import_QESU_PHY_v3}
\title{import_QESU_PHY_v3}
\usage{
import_QESU_PHY_v3(x, complet = FALSE)
}
\arguments{
\item{x}{un fichier xml conforme au scénario d'échange QUESU_PHY_v3 du SANDRE}

\item{complet}{booleen FALSE par défaut. S'il est FALSE alors le fichier résultat comporte seulement les informations essentielles. S'il est TRUE, l'ensemble des colonnes sont conservées.}
}
\value{
la fonction renvoie une liste de data.frame avec les résultats contenus dans le fichier
}
\description{
fonction pour importer un objet xml de type QUESU_PHY_V3 sous forme d'objet R
}
\examples{
x<-"C:\\\\Users\\\\anthony.deburghrave\\\\OneDrive - EPTB Vilaine\\\\Documents\\\\R_Anthony\\\\tools4DCE\\\\exemples xml\\\\QUESU3_EPTB_VILAINE_2020-11-04 v2.xml"

donnees<-import_QESU_PHY_v3(x)
}
