% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend_LQ.R
\name{legend_LQ}
\alias{legend_LQ}
\title{legend_LQ}
\usage{
legend_LQ(taille_legende = 10)
}
\arguments{
\item{taille_legende}{taille de la police de la légende}
}
\value{
la fonction renvoie un objet cowplot avec la légende
}
\description{
fonction pour afficher un objet cowplot avec la légende des graphiques DCE pour les valeurs inférieuresà la LQ
}
\examples{
plot_grid(legend_LQ())

}
