% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSeuils.R
\name{makeSeuils}
\alias{makeSeuils}
\title{makeSeuils}
\usage{
makeSeuils()
}
\value{
la fonction renvoie une liste d'objets de class seuil
}
\description{
Créé automatiquement une liste de seuils à partir des données stockées dans les objets base_seuils, couleurs_classes et ordre_facteurs_qualite
}
\examples{
makeSeuils()
}
