% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordre_facteurs_qualite.R
\docType{data}
\name{ordre_facteurs_qualite}
\alias{ordre_facteurs_qualite}
\title{ordre_facteurs_qualite}
\format{
An object of class \code{data.frame} with 49 rows and 1 columns.
}
\usage{
data(ordre_facteurs_qualite)
}
\description{
data frame contenant l'ordre à prendre en compte dans les "levels" des classes de seuil.
Permet de respecter l'ordre de la meilleure qualité vers la pire
}
\details{
La colonne CLASSE renvoie aux classes de la base_seuils
}
\examples{
head(data(ordre_facteurs_qualite))
}
\keyword{datasets}
