% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametres_sandre.R
\docType{data}
\name{parametres_sandre}
\alias{parametres_sandre}
\title{parametres_sandre}
\format{
An object of class \code{data.frame} with 6066 rows and 5 columns.
}
\source{
\url{https://api.sandre.eaufrance.fr/referentiels/v1/par.csv?outputSchema=SANDREv4&compress=true}
}
\usage{
data(parametres_sandre)
}
\description{
data frame contenant le référentiel SANDRE paramètre.
}
\examples{
head(data(parametres_sandre))

}
\keyword{datasets}
