% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentileDCE.R
\name{PercentileDCE}
\alias{PercentileDCE}
\title{PercentileDCE}
\usage{
PercentileDCE(A, type = 0.9, na.rm = T)
}
\arguments{
\item{A}{= vecteur de nombre dans lequel il faut calculer le percentile}

\item{type}{= type de percentile (0.9 pour percentile 90, 0.1 pour percentile 10)}

\item{na.rm}{= paramètre pour ne pas prendre en compte les valeurs non numériques (TRUE) ou renvoyer une erreur (F)}
}
\value{
le percentile calculé selon les règles DCE françaises
}
\description{
fonction pour calculer le percentile DCE selon les règles françaises d'évaluation de l'état écologique.
Ce percentile se calcule selon la loi de Hazen (il s’agit de la valeur du rang égal à l’arrondi du produit de 0,9 par le nombre de valeur auquel on ajoute 0,5.)
}
\details{
https://www.artois-picardie.eaufrance.fr/qualite-et-quantite-des-eaux/reseaux-de-surveillance/article/5eme-campagne-de-surveillance
}
\examples{
test<-sample(x=1:100, size=20, replace=T)
PercentileDCE(test, type=0.9)
}
