% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_taxo_sandre.R
\docType{data}
\name{ref_taxo_sandre}
\alias{ref_taxo_sandre}
\title{ref_taxo_sandre}
\format{
An object of class \code{data.frame} with 68796 rows and 18 columns.
}
\usage{
data(ref_taxo_sandre)
}
\description{
data frame contenant le référntiel taxonomique du SANDRE
https://www.sandre.eaufrance.fr/atlas/srv/eng/catalog.search#/metadata/0bea2eb9-a4ee-4a81-b494-bd4fb5338ef8
}
\examples{
str(data(ref_taxo_sandre))
}
\keyword{datasets}
