% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuils_nonDCE.R
\docType{data}
\name{seuils_nonDCE}
\alias{seuils_nonDCE}
\title{seuils_nonDCE}
\format{
An object of class \code{character} of length 5.
}
\usage{
data(seuils_nonDCE)
}
\description{
vecteur comportant 5 couleurs utilisé pour les fonds colorés des graphiques représentant des paramètres non DCE
}
\examples{
data(seuils_nonDCE)
}
\keyword{datasets}
