% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuils_Tb2B2Mo2Me2Ma2.R
\docType{data}
\name{seuils_Tb2B2Mo2Me2Ma2}
\alias{seuils_Tb2B2Mo2Me2Ma2}
\title{seuils_Tb2B2Mo2Me2Ma2}
\format{
An object of class \code{character} of length 10.
}
\usage{
data(seuils_Tb2B2Mo2Me2Ma2)
}
\description{
vecteur comportant 10 couleurs utilisé pour les fonds colorés des graphiques représentant des paramètres DCE à 5 seuils de type pH
}
\examples{
data(seuils_Tb2B2Mo2Me2Ma2)
}
\keyword{datasets}
