% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuils_TbBMo.R
\docType{data}
\name{seuils_TbBMo}
\alias{seuils_TbBMo}
\title{seuils_TbBMo}
\format{
An object of class \code{character} of length 3.
}
\usage{
data(seuils_TbBMo)
}
\description{
vecteur comportant 3 couleurs utilisé pour les fonds colorés des graphiques représentant des paramètres DCE à 3 seuils comme les polluants spécifiques de l'état écologique
}
\examples{
data(seuils_TbBMo)
}
\keyword{datasets}
