% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuils_TbBMoMeMa.R
\docType{data}
\name{seuils_TbBMoMeMa}
\alias{seuils_TbBMoMeMa}
\title{seuils_TbBMoMeMa}
\format{
An object of class \code{character} of length 5.
}
\usage{
data(seuils_TbBMoMeMa)
}
\description{
vecteur comportant 5 couleurs utilisé pour les fonds colorés des graphiques représentant des paramètres DCE à 5 seuils de type PC classique
}
\examples{
data(seuils_TbBMoMeMa)
}
\keyword{datasets}
