% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_distribution.R
\name{table_distribution}
\alias{table_distribution}
\title{table_distribution}
\usage{
table_distribution(donnees, col_Valeur = "RsAna", col_CdRq = "CdRqAna", seuil)
}
\arguments{
\item{donnees}{: dataframe contenant les analyses dont il faut faire le bilan}

\item{col_Valeur}{: nom de la colonne avec les résultats d'analyses (par défaut RsAna)}

\item{col_CdRq}{: nom de la colonne avec le code remarque des analyses (par défaut CdRqAna). Les codes correspondent à ceux définis par le SANDRE.}

\item{seuil}{: object de classe seuil ou liste d'objets de classe seuil}

\item{col_parametre}{: nom de la colonne par rapport à laquelle faire le bilan (par défaut NomParametre)}
}
\value{
la fonction renvoie une data.frame avec le nombre d'analyses dans chaque classe et le nombre de données <LQ dans chaque classe
}
\description{
fonction pour faire le bilan de la distribution des données par rapport à un seuil
Ce tableau fait le bilan du nombre de données disponibles par classe de qualité
en distinguant les valeurs quantifiées de celles <LQ
}
\examples{
donnees<-data.frame(RsAna=sample(0.1:60, 500, replace=TRUE), LqAna=c(0.5,1,2,12))
donnees<-donnees\%>\%mutate(RsAna=ifelse(RsAna<LqAna, LqAna, RsAna))
donnees<-donnees\%>\%mutate(CdRqAna=ifelse(RsAna>LqAna, "1", ifelse(sample(1:100,5)>10,"10","1")))
seuil<-makeSeuils(CdParametre="1340", type_seuil="DCE")
table_distribution(donnees, seuil)
}
