% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tri_molecules.R
\name{tri_molecules}
\alias{tri_molecules}
\title{tri_molecules}
\usage{
tri_molecules(donnees, tri_croissant = TRUE, ordre_facteurs = NULL)
}
\arguments{
\item{donnees}{un data.frame avec les colonnes obtenues en sortie de la fonction groupe_tableau_distribution}

\item{tri_croissant}{booleen vrai par défaut. Si faux, l'ordre du tri est inversé.}

\item{ordre_facteurs}{: vecteur optionnel qui ordonne les levels si on ne veut pas retenir le classement par défaut}
}
\value{
la fonction renvoie un vecteur contenant les paramètres les plus impactés. Le tri se fait sur la base suivant : en 1er est renvoyé le parametre qui a été quantifié le plus souvent au niveau le plus dégradé (en pourcentage de quantification sur nb total_tri de mesures)
}
\description{
fonction pour trier les x molécules les plus préoccupantes parmi un ensemble de résultats
recoit en entree un tableau de distribution genere par la fonction groupe_tableau_distribution
}
\examples{
donnees<-data.frame(parametres=rep(c("1301", "1340", "1335"), 100), RsAna=sample(0.1:100, 300, replace=TRUE), LqAna=c(0.5,1,6))
donnees<-donnees\%>\%mutate(RsAna=ifelse(RsAna<LqAna, LqAna, RsAna))
donnees<-donnees\%>\%mutate(CdRqAna=ifelse(RsAna>LqAna, "1", ifelse(sample(1:100,5)>10,"10","1")))
seuils<-makeSeuils(CdParametre=donnees$parametres\%>\%unique, specificites=c("CYPRINICOLE", rep(NA,2)), type_seuil = "DCE")
donnees<-groupe_tableau_distribution(donnees, col_CdParametre="parametres", col_CdSupport=NULL, col_CdFraction=NULL, col_CdUnite=NULL, seuils = seuils)
tri_molecules(donnees)
}
