% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajoute_nom_cdqualana.R
\name{ajoute_nom_cdqualana}
\alias{ajoute_nom_cdqualana}
\title{ajoute_nom_cdqualana}
\usage{
ajoute_nom_cdqualana(x, col_qualif = "CdQualAna")
}
\arguments{
\item{x}{data.frame avec le code paramètre}

\item{col_qualif}{: nom de la colonne avec le code qualification Valeur par défaut : CdQualAna}
}
\value{
la fonction le data.frame complété de la colonne RqAna
}
\description{
fonction pour ajouter le nom du code qualification de l'analyse sandre à une data frame
}
\examples{
test<-data.frame(CdQualAna=c("1", "10"))
ajoute_nom_cdqualana(test, col_qualif="CdQualAna")
}
