% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcule_EQR_hbio.R
\name{calcule_EQR_hbio}
\alias{calcule_EQR_hbio}
\title{calcule_EQR_hbio}
\usage{
calcule_EQR_hbio(
  valeur_indice,
  CdParametre,
  typologie,
  specificite = "CAS_GENERAL"
)
}
\arguments{
\item{valeur_indice}{note à convertir en EQR}

\item{CdParametre}{code sandre du paramètre à convertir}

\item{typologie}{: typologie de la station à convertir}

\item{specificite}{: précisions sur la typologie (ex pour IBD de > 10000 km2 ou pas)}
}
\value{
la fonction renvoie l'AQR correspondant à la note.
}
\description{
fonction pour calculer l'EQR d'un indice hydrobiologique à partir de la note de l'indice et de la typologie de la staation.
}
\examples{
calcule_EQR_hbio(15.5, CdParametre="5856", typologie="P12-A", specificite="CAS_GENERAL")
}
