% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charge_shp_ICPE.R
\name{charge_shp_ICPE}
\alias{charge_shp_ICPE}
\title{charge_shp_ICPE}
\usage{
charge_shp_ICPE(crs = 2154, shp_emprise = NULL, rubrique = TRUE)
}
\arguments{
\item{crs}{valeur du code de projection dans lequel renvoyer le résultat (par défaut Lambert 93, indiquer 4326 pour du wgs84)}

\item{shp_emprise}{vecteur (optionnel) objet sf dans lequel on va récupérer les ICPE}

\item{rubrique}{booléen : si vrai (valeur par défaut), le script appelle chaque page internet d'ICPE pour récupérer les rubriques de classement.}
}
\value{
la fonction renvoie une liste avec un slot shp = objet sf avec les ICPE de l'emprise indiquée,

avec un slot rejets = table avec les rejets référencés dans le registre national des émissions polluantes,

et un slot prelevements avec les prélèvements d'eau issus de ce registre
}
\description{
charge sous forme d'objet le shape des ICPE depuis https://www.georisques.gouv.fr/donnees/bases-de-donnees/installations-industrielles
et les émissions polluantes depuis https://www.georisques.gouv.fr/donnees/bases-de-donnees/installations-industrielles-rejetant-des-polluants
}
\examples{
ICPE<-charge_shp_ICPE(shp_emprise=charge_shp_SAGE(nom_sage="Vilaine"))
}
