% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtre_donnees_extremes.R
\name{filtre_donnees_extremes}
\alias{filtre_donnees_extremes}
\title{filtre_donnees_extremes}
\usage{
filtre_donnees_extremes(x, xmin = NA, xmax = NA)
}
\arguments{
\item{x}{un vecteur numerique}

\item{xmin}{un nombre}

\item{xmax}{un nombre}
}
\value{
la fonction renvoie un vescteur avec les données extrèmes remplacées par ymin ou ymax.
}
\description{
filtre les donnees extremes qui n'apparaitront pas parmi les points des graphs
pour cela on ecrete les valeurs entrant en remplacant les valeurs inférieures à xmin ou supérieures à xmax par xmin ou xmax
}
\examples{
filtre_donnees_extremes(seq(1:10), xmin=3, xmax=9)
}
