% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fractions_sandre.R
\docType{data}
\name{fractions_sandre}
\alias{fractions_sandre}
\title{fractions_sandre}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 330 rows and 2 columns.
}
\usage{
data(fractions_sandre)
}
\description{
data frame contenant les correspondances codes SANDRE de fraction et son nom
jeu de données téléchargé depuis https://www.sandre.eaufrance.fr/
}
\examples{
head(data(fractions_sandre))
}
\keyword{datasets}
