% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphDCE_distribution.R
\name{graphDCE_distribution}
\alias{graphDCE_distribution}
\title{graphDCE_distribution}
\usage{
graphDCE_distribution(
  donnees,
  titre = "",
  taille_titre = 12,
  taille_axes = 11,
  legende_LQ = c("NON QUANTIFIE", "QUANTIFIE"),
  affiche_valeurs = T,
  min_affiche = 0.1,
  tri_donnees = T,
  tri_croissant = T,
  nb_top = NULL
)
}
\arguments{
\item{donnees}{: dataframe contenant les analyses dont il faut faire le bilan dans un format du type de celui rendu par la fonction groupe_tableau_distribution}

\item{titre}{: titre du graphique}

\item{taille_titre}{taille police du titre (par défaut 12)}

\item{taille_axes}{taille police des axes (par défaut 11)}

\item{legende_LQ}{: vecteur de character de longueur 2 qui précise l'étiquette pour les données inférieures et supérieures à la LQ}

\item{affiche_valeurs}{: booléen qui indique d'il faut afficher les valeurs ou pas dans les barres}

\item{min_affiche}{: pourcentage minimum à atteindre pour que la valeur soit affiché (ex si min_affiche=0.3, seules le nb de prélèvements représentant au moins 30\% pour une categorie et un paramètre sera affiché)}

\item{tri_donnees}{booleen vrai par défaut. Si il est vrai, les résultats sont affichés par ordre du plus dégradé au moins dégradé (au sens des dépassement de seuils). Si non, les résultats sont affichés dans l'ordre alphabétique des noms de paramètres.}

\item{tri_croissant}{booleen vrai par défaut. Si faux, l'ordre du tri est inversé. Paramètre sans effet si tri_donnees=FALSE}

\item{nb_top}{: optionnel. Si renseigné, le graph n'affiche que le nb de paramètres indiqués en affichant en priorité les + dégradés}
}
\value{
la fonction renvoie un objet ggplot avec le graphique de distribution
}
\description{
fonction pour réaliser un graphique bilan de quantifications
}
\examples{
donnees<-data.frame(parametres=rep(c("1301", "1340", "1335"), 100), RsAna=sample(0.1:100, 300, replace=TRUE), LqAna=c(0.5,1,6))
donnees<-donnees\%>\%mutate(RsAna=ifelse(RsAna<LqAna, LqAna, RsAna))
donnees<-donnees\%>\%mutate(CdRqAna=ifelse(RsAna>LqAna, "1", ifelse(sample(1:100,5)>10,"10","1")))
seuils<-makeSeuils(CdParametre=donnees$parametres\%>\%unique, specificites=c("CYPRINICOLE", rep(NA,2)), type_seuil = "DCE")
tableau<-groupe_tableau_distribution(donnees, col_CdParametre="parametres", col_CdSupport=NULL, col_CdFraction=NULL, col_CdUnite=NULL, seuils = seuils)
graphDCE_distribution(tableau)
}
