% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P90glissant.R
\name{P90glissant}
\alias{P90glissant}
\title{P90glissant}
\usage{
P90glissant(analyses, col_date = "DatePrel", col_rsana = "RsAna", nb_mois)
}
\arguments{
\item{analyses}{= dataframe de données}

\item{col_date}{= nom de la colonne avec la date de prélèvement, par défaut Dateprel}

\item{col_rsana}{= nom de la colonne avec les résultats d'analyses, par défaut RsAna}

\item{nb_mois}{= nombre de mois sur lesquels faire le calcul}
}
\value{
la fonction renvoie un dataframe avec une colonne date avec les mois et une colonne P90 avec
le percentile DCE 90 sur la période entre la date et la date - nb_mois
}
\description{
fonction pour calculer le percentile 90 glissant sur un nombre de mois donnés
}
\examples{
set.seed(123)

df <- data.frame(
  DatePrel = sample(
    seq(
      as.Date("2015-01-01"),
      as.Date("2023-12-31"),
      by = "day"
    ),
    500
  ),
  RsAna = runif(500, 0, 100)
) |>
  dplyr::arrange(DatePrel)

P90glissant(df, nb_mois = 36)
}
