% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setSeuil.R
\name{setSeuils}
\alias{setSeuils}
\title{setSeuils}
\usage{
setSeuils(
  nom_parametre,
  nom_seuil,
  type_seuil,
  code_parametre,
  support = "",
  fraction = "",
  code_unite = "",
  seuils,
  bornesinfinclue = T,
  specificites = ""
)
}
\arguments{
\item{nom_parametre}{= un character qui identifie explicitement le paramètre représenté}

\item{nom_seuil}{= un character qui précise l'origine des seuils}

\item{type_seuil}{= un character qui désigne type du seuil (DCE ou autre)}

\item{code_parametre}{= Code SANDRE du paramètre (optionnel)}

\item{support}{= Code SANDRE du support (optionnel)}

\item{fraction}{= Code SANDRE de la fraction (optionnel)}

\item{code_unite}{= Code SANDRE de l'unité (optionnel)}

\item{seuils}{= data.frame avec les colonnes suivantes : CLASSE (character), SEUILMIN (numeric), SEUILMAX (numeric), NOM_COULEUR (un nom ou code hexa de couleur valide)}

\item{bornesinfinclue}{= booléen. Si vrai la classe de qualité couvrira l'intervalle [SEUILMIN, SEUILMAX\link{, si faux elle couvrira }SEUIL_MIN,SEUIL_MAX]}

\item{specificites}{= typologie particulière à laquelle s'applique le seuil}
}
\value{
la fonction renvoie un objet de class seuil
}
\description{
création d'un objet de classe seuil
}
\examples{
setSeuils(nom_parametre="parametre test",nom_seuil="AM 25 janv 2010",type_seuil="DCE", code_parametre="1301",support="3",code_unite="5", seuils=tools4DCE::base_seuils\%>\%subset(NOM=="TEMPERATURE" & SPECIFICITE=="CYPRINICOLE")\%>\%left_join(couleurs_classes, by=c("CLASSE", "TYPE")),bornesinfinclue=T, levels_classes=c("TRES BON", "BON", "MOYEN","MEDIOCRE", "MAUVAIS"))

}
