% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ref_eqr.R, R/class_seuil.R,
%   R/class_station.R
\docType{class}
\name{seuil-class}
\alias{seuil-class}
\title{class Seuil}
\description{
Définition des objets de classe seuil

Définition des objets de classe seuil

Définition des objets de classe seuil
}
\details{
Ces objets précisent les limites de chaque classe de qualité

accessors :

nom_parametre = Nom du paramètre à représenter graphiquement
nom_seuil= Origine du seuil (arrêté ministériel du 25 janvier 2010, ...)
type_seuil = type du seuil (DCE ou autre)
code_parametre = Code SANDRE du paramètre
synonymes_parametre = Codes SANDRE du paramètre + de ses synonymes. Les codes sont séparés par |
support = Code SANDRE du support (optionnel)
fraction = Code SANDRE de la fraction (optionnel)
code_unite = Code SANDRE de l'unité (optionnel)
seuils = data.frame avec les colonnes suivantes : CLASSE (character), SEUILMIN (numeric), SEUILMAX (numeric), NOM_COULEUR (un nom ou code hexa de couleur valide)
bornesinfinclue = booléen. Si vrai la classe de qualité couvrira l'intervalle [SEUILMIN, SEUILMAX\link{, si faux elle couvrira }SEUIL_MIN,SEUIL_MAX]
specificites = typologie particulière à laquelle s'applique le seuil

Ces objets précisent les limites de chaque classe de qualité

accessors :

nom_parametre = Nom du paramètre à représenter graphiquement
nom_seuil= Origine du seuil (arrêté ministériel du 25 janvier 2010, ...)
type_seuil = type du seuil (DCE ou autre)
code_parametre = Code SANDRE du paramètre
synonymes_parametre = Codes SANDRE du paramètre + de ses synonymes. Les codes sont séparés par |
support = Code SANDRE du support (optionnel)
fraction = Code SANDRE de la fraction (optionnel)
code_unite = Code SANDRE de l'unité (optionnel)
seuils = data.frame avec les colonnes suivantes : CLASSE (character), SEUILMIN (numeric), SEUILMAX (numeric), NOM_COULEUR (un nom ou code hexa de couleur valide)
bornesinfinclue = booléen. Si vrai la classe de qualité couvrira l'intervalle [SEUILMIN, SEUILMAX\link{, si faux elle couvrira }SEUIL_MIN,SEUIL_MAX]
specificites = typologie particulière à laquelle s'applique le seuil

Ces objets précisent les limites de chaque classe de qualité

accessors :

nom_parametre = Nom du paramètre à représenter graphiquement
nom_seuil= Origine du seuil (arrêté ministériel du 25 janvier 2010, ...)
type_seuil = type du seuil (DCE ou autre)
code_parametre = Code SANDRE du paramètre
synonymes_parametre = Codes SANDRE du paramètre + de ses synonymes. Les codes sont séparés par |
support = Code SANDRE du support (optionnel)
fraction = Code SANDRE de la fraction (optionnel)
code_unite = Code SANDRE de l'unité (optionnel)
seuils = data.frame avec les colonnes suivantes : CLASSE (character), SEUILMIN (numeric), SEUILMAX (numeric), NOM_COULEUR (un nom ou code hexa de couleur valide)
bornesinfinclue = booléen. Si vrai la classe de qualité couvrira l'intervalle [SEUILMIN, SEUILMAX\link{, si faux elle couvrira }SEUIL_MIN,SEUIL_MAX]
specificites = typologie particulière à laquelle s'applique le seuil
}
