% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stations.R
\docType{data}
\name{stations}
\alias{stations}
\title{stations}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 37572 rows and 2 columns.
}
\usage{
data(stations)
}
\description{
data frame contenant les correspondances codes SANDRE des stations de mesures et leur nom sandre
jeu de données téléchargé avec la fonction charge_shp_STAQ() du package
}
\examples{
head(data(stations))
}
\keyword{datasets}
