% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supports_sandre.R
\docType{data}
\name{supports_sandre}
\alias{supports_sandre}
\title{supports_sandre}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 114 rows and 2 columns.
}
\usage{
data(supports_sandre)
}
\description{
data frame contenant les correspondances codes SANDRE du support et son nom
jeu de données téléchargé depuis https://www.sandre.eaufrance.fr/
}
\examples{
head(data(supports_sandre))
}
\keyword{datasets}
