% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unites_sandre.R
\docType{data}
\name{unites_sandre}
\alias{unites_sandre}
\title{unites_sandre}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 645 rows and 11 columns.
}
\usage{
data(unites_sandre)
}
\description{
data frame contenant les correspondances codes SANDRE de l'unité - symbole de l'unité
jeu de données téléchargé depuis https://www.sandre.eaufrance.fr/atlas/srv/fre/catalog.search#/metadata/dbae73ac-7118-4424-866e-c9a0980ee519
}
\examples{
head(data(unites_sandre))
}
\keyword{datasets}
